app-service/requirements.txt
flasgger==0.9.7.1
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
python-dotenv==1.1.0
requests==2.32.3